fix(api): ensure user-settings are created for new users
The problem was that in the request definition we had the
`custom_output_bucket` field as required, but in this case when the
user is creating it that field doesn't exist.

After setting as optional the request continues, and doesn't really
matter to add a fallback value because when creating it, in the
database we set that column to FALSE by default.