fix(workspace): include gpu_event in workspace-triggered runs
This is needed because we need to track the runs when they are attached
to a session.

So makes sense to add the gpu_event when storing the run in the
database. You can queue 20 runs and all those are supposed to be
executed in the same session, if the session finish all those runs
queued should be canceled.