fix(workflows): update only runs that are not in a finish state
since modal call this endpoint to update the status of a run, we can
finish with a race condition where modal sends this request to update a
run when this was already marked as cancel in other process.

This is a problem because overrides the cancel or any end state
`["success", "failed", "timeout", "cancelled"]`

To avoid this I'm adding a condition in the update query to only update
those records that doesn't have an end state.

previously:
```
update_stmt = (
            update(WorkflowRun)
            .where(
                WorkflowRun.id == body.run_id,
            )
            .values(**update_values)
        )
```

Now:
```
update_stmt = (
            update(WorkflowRun)
            .where(
                and_(
                    WorkflowRun.id == body.run_id,
                    ~WorkflowRun.status.in_(endStatuses)
                )
            )
            .values(**update_values)
        )
```