fix(workflows): display cosmos video outputs
The problem is that the the validation of the output against the
definition doesn't match at all.
`WorkflowRunOutputModel.model_validate(output)`

In the output saved we have something like this
```
{
	"images": ...,
	"animated": bool
}
˜``
As you can see we have an animated property that has a bool value, and
the problem is that in the definition we only expect to this value be
a string.