Dev (#185)
* fix(workspace): commit db changes when gpu_event is gpu_end

* logs

* Revert "logs"

This reverts commit 74084234c9e701ca2a509f46313f4c57e278ebe3.

* fix(workspace): create a session in old v4 machines

old v4 machines doesn't have increase timeout method.
Previously was working with `hasattr(runner, 'increase_timeout')`, but
it's possible that with the update of modal cli or an update on their
end, this little hack doesn't work anymore.

* feat!: dynamic session + session creation will return faster

* update: extra model paths

* fix: creating session from merge

* fix: starting session with machine

* fix(workflows): display cosmos video outputs

The problem is that the the validation of the output against the
definition doesn't match at all.
`WorkflowRunOutputModel.model_validate(output)`

In the output saved we have something like this
```
{
	"images": ...,
	"animated": bool
}
˜``
As you can see we have an animated property that has a bool value, and
the problem is that in the definition we only expect to this value be
a string.

* feat: modal image id support for dynamic session

---------

Co-authored-by: EmmanuelMr18 <emm_mr@hotmail.com>
Co-authored-by: nick <kobenkao@gmail.com>