fix(storage): return 409 on duplicated model request
This was returning 400 code but then getting wrapped with 500 code,
and that is why we were having the logs in logfire that are just
creating noise.

They are creating just noise because a duplicated model is an error
from the user (that is why we return now 409 - Conflict). We cannot
really do something with that log, there is not an action item