Dev (#192)
* fix(workspace): commit db changes when gpu_event is gpu_end

* logs

* Revert "logs"

This reverts commit 74084234c9e701ca2a509f46313f4c57e278ebe3.

* fix(workspace): create a session in old v4 machines

old v4 machines doesn't have increase timeout method.
Previously was working with `hasattr(runner, 'increase_timeout')`, but
it's possible that with the update of modal cli or an update on their
end, this little hack doesn't work anymore.

* feat!: dynamic session + session creation will return faster

* update: extra model paths

* fix: creating session from merge

* fix: starting session with machine

* fix(workflows): display cosmos video outputs

The problem is that the the validation of the output against the
definition doesn't match at all.
`WorkflowRunOutputModel.model_validate(output)`

In the output saved we have something like this
```
{
	"images": ...,
	"animated": bool
}
˜``
As you can see we have an animated property that has a bool value, and
the problem is that in the definition we only expect to this value be
a string.

* feat: modal image id support for dynamic session

* fix: dyanmic sessions

* feat: support modal image id in builder

* fix: deployment

* docs(blog): how to install IDM VTON custom nodes

https://github.com/TemryL/ComfyUI-IDM-VTON/

* docs(blog): how to install IDM VTON custom nodes, update titles

* feat: remove legazy machine url

* clean up v4 app, removing api

* fix: too many values to unpack (expected 2) for getting workflow version

https://discord.com/channels/1185598676983349338/1317310131368755200/1331517524495437888

* feat: retry 3 times every 5 second interval for runpod

* test: add fake fail 500 for testing

* Revert "test: add fake fail 500 for testing"

This reverts commit a2d45a16ac61e6d0ba16615b920cea17159c92eb.

* tweak: revert back

* fix: retry to 2 seconds

* feat: start dynamic session should have comfyui manager

* fix: make the retry to exponential

* fix(storage): return 409 on duplicated model request

This was returning 400 code but then getting wrapped with 500 code,
and that is why we were having the logs in logfire that are just
creating noise.

They are creating just noise because a duplicated model is an error
from the user (that is why we return now 409 - Conflict). We cannot
really do something with that log, there is not an action item

* chore: update dynamic session creation

* feat: redirect to new blog

* fix: optional body for snapshto

* fix: single db execute to get version

---------

Co-authored-by: EmmanuelMr18 <emm_mr@hotmail.com>
Co-authored-by: nick <kobenkao@gmail.com>
Co-authored-by: karrix <thomasdice1834793903@gmail.com>